requirements/base.txt
# Bleeding edge Django
django==1.8.5

# Configuration
django-environ==0.4.0
django-secure==1.0.1

# Forms
django-braces==1.8.1
django-crispy-forms==1.5.2
django-floppyforms==1.5.2

# Models
django-model-utils==2.3.1

# Images
Pillow==3.0.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.23.0

# For the persistence stores
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1
django-autoslug==1.8.0

# Time zones support
pytz==2015.6

# Redis support
django-redis==4.2.0
redis>=2.10.0

# Your custom requirements go here
ShopifyAPI==2.1.5
runstats==0.5.3

# Loginas
django-loginas==0.1.5

iso8601==0.1.10

django-bootstrap3==6.2.2

# Email
djrill==2.0

#django-plans
django-plans==0.8.11

#django-payments
django-payments==0.9.1
stripe==1.27.1

#django cron
django-cron==0.4.6
